fix: 增加orm钩子，不填uuid时自动生成一个uuid

+++ main/app/repository/base/gift_or_free_order_reason.go
@@ -9,10 +9,10 @@ import (
 
 // IGiftOrFreeOrderReasonRepo 赠品或免费订单原因仓库接口
 type IGiftOrFreeOrderReasonRepo interface {
-	GetGiftOrFreeOrderReasonList() ([]model.FreeReason, error)                         // 获取赠品或免费订单原因列表
-	UpdateGiftOrFreeOrderReason(id uint, giftOrFreeOrderReason model.FreeReason) error // 更新赠品或免费订单原因
-	CreateGiftOrFreeOrderReason(giftOrFreeOrderReason model.FreeReason) (uint, error)  // 创建赠品或免费订单原因
-	DeleteGiftOrFreeOrderReason(id uint) error                                         // 删除赠品或免费订单原因
+	GetGiftOrFreeOrderReasonList() ([]model.FreeReason, error)                           // 获取赠品或免费订单原因列表
+	UpdateGiftOrFreeOrderReason(id uint64, giftOrFreeOrderReason model.FreeReason) error // 更新赠品或免费订单原因
+	CreateGiftOrFreeOrderReason(giftOrFreeOrderReason model.FreeReason) (uint64, error)  // 创建赠品或免费订单原因
+	DeleteGiftOrFreeOrderReason(id uint64) error                                         // 删除赠品或免费订单原因
 }
 
 // NewGiftOrFreeOrderReasonRepo 创建新的赠品或免费订单原因仓库
@@ -37,7 +37,7 @@ func (r *GiftOrFreeOrderReasonRepoImpl) GetGiftOrFreeOrderReasonList() ([]model.
 }
 
 // UpdateGiftOrFreeOrderReason 更新赠品或免费订单原因
-func (r *GiftOrFreeOrderReasonRepoImpl) UpdateGiftOrFreeOrderReason(id uint, giftOrFreeOrderReason model.FreeReason) error {
+func (r *GiftOrFreeOrderReasonRepoImpl) UpdateGiftOrFreeOrderReason(id uint64, giftOrFreeOrderReason model.FreeReason) error {
 	tx := r.db.Begin() // 开始事务
 	defer func() {
 		if r := recover(); r != nil {
@@ -59,7 +59,7 @@ func (r *GiftOrFreeOrderReasonRepoImpl) UpdateGiftOrFreeOrderReason(id uint, gif
 }
 
 // CreateGiftOrFreeOrderReason 创建赠品或免费订单原因
-func (r *GiftOrFreeOrderReasonRepoImpl) CreateGiftOrFreeOrderReason(giftOrFreeOrderReason model.FreeReason) (uint, error) {
+func (r *GiftOrFreeOrderReasonRepoImpl) CreateGiftOrFreeOrderReason(giftOrFreeOrderReason model.FreeReason) (uint64, error) {
 	tx := r.db.Begin() // 开始事务
 	defer func() {
 		if r := recover(); r != nil {
@@ -83,6 +83,6 @@ func (r *GiftOrFreeOrderReasonRepoImpl) CreateGiftOrFreeOrderReason(giftOrFreeOr
 }
 
 // DeleteGiftOrFreeOrderReason 软删除赠品或免费订单原因
-func (r *GiftOrFreeOrderReasonRepoImpl) DeleteGiftOrFreeOrderReason(id uint) error {
+func (r *GiftOrFreeOrderReasonRepoImpl) DeleteGiftOrFreeOrderReason(id uint64) error {
 	return r.db.Model(&model.FreeReason{}).Where("id = ?", id).Update("delete_time", uint(time.Now().Unix())).Error
 }
